修改最小金额为浮点型

+++ admin/api/cards.py
@@ -338,7 +338,7 @@ class ApiCardBuy(Resource):
             price = args.get('price')
             count = args.get('count')
             openid = args.get('openId')
-            order = create_order(card_id, int(price), openid, count)
+            order = create_order(card_id, float(price), openid, count)
             if not order:
                 return {'result': 250}
             res, outputs = payable(request, openid, order)
